renamed feeding-buttons and updated imports
